pyproject.toml
[project]
name = "redesneuronales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.13.0",
    "duckduckgo-search>=8.1.1",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.3",
]
